Add user profiles and search post

+++ src/main/java/com/afect/controller/UserController.java
@@ -2,6 +2,7 @@ package com.afect.controller;
 
 import java.util.List;
 
+import org.apache.log4j.Logger;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
@@ -25,18 +26,25 @@ import com.afect.service.UserService;
 @CrossOrigin(origins = "*")
 public class UserController {
 	private UserService uService;
+	private static final Logger LOGGER = Logger.getLogger(UserController.class);
 	
 	public UserController()
 	{
 		super();
+		LOGGER.info("Init User Controller");
 	}
 
 	@Autowired
 	public UserController(UserService uService) {
 		super();
 		this.uService = uService;
+		LOGGER.info("Auto Init User Controller");
 	}
 	
+	/*
+	 * Checks if an email is not already associated with an account
+	 * Called when user is filling out the Registration form in /Register
+	 * */
 	@PostMapping("/checkEmail/{email}")
 	public ResponseEntity<String> checkEmail(@PathVariable("email") String email)
 	{
@@ -45,6 +53,7 @@ public class UserController {
 		//If no user is found, then email is available
 		if(uService.getUserByEmail(email) == null)
 		{
+			LOGGER.trace("Email: " + email + " is available");
 			return new ResponseEntity<String>("Email is available", HttpStatus.OK);
 		}
 		
@@ -52,12 +61,17 @@ public class UserController {
 		return new ResponseEntity<String>("Email is taken", HttpStatus.FOUND);
 	}
 	
+	/*
+	 * Check is a username is available to use
+	 * Called when user is filling out the Registration form in /Register
+	 * */
 	@PostMapping("/checkUsername/{un}")
 	public ResponseEntity<String> checkUsername(@PathVariable("un") String un)
 	{
 		//If no user is found, then email is available
 		if(uService.getUserByUsername(un) == null)
 		{
+			LOGGER.trace("Username: " + un + " is available");
 			return new ResponseEntity<String>("un is available", HttpStatus.OK);
 		}
 		
@@ -67,6 +81,7 @@ public class UserController {
 	
 	/*
 	 * Attempt to register a user
+	 * Called when user hits the Register button in /Register
 	 * */
 	@PostMapping
 	public ResponseEntity<String> addUser(@RequestBody User u)
@@ -76,13 +91,13 @@ public class UserController {
 		//First try and find this user by email
 		if(!(uService.getUserByEmail(u.getEmail()) == null))
 		{
-			System.out.println("User email already exists");
+			LOGGER.warn("Attempted to register with an existing email");
 			return new ResponseEntity<String>("Failed to create user", HttpStatus.BAD_REQUEST);
 		}
 		//Try and find by username
 		else if(!(uService.getUserByEmail(u.getUsername()) == null))
 		{
-			System.out.println("Username already exists");
+			LOGGER.warn("Attempted to register with an existing username");
 			return new ResponseEntity<String>("Failed to create user", HttpStatus.BAD_REQUEST);
 		}
 		
@@ -93,10 +108,12 @@ public class UserController {
 		}
 		catch(Exception e)
 		{
-			System.out.print(e.getMessage());
+			e.printStackTrace();
+			LOGGER.error(e.getMessage());
 			return new ResponseEntity<String>("Failed to create user", HttpStatus.BAD_REQUEST);
 		}
 		
+		LOGGER.info("New user created: " + u);
 		return new ResponseEntity<String>("Resource was created", HttpStatus.CREATED);
 	}
 	/*
@@ -113,16 +130,19 @@ public class UserController {
 		//If no user found OR password does not match, Return 404
 		if (u == null || !u.getPassword().equals(login.getPassword()))
 		{
-			System.out.println("No user found");
+			LOGGER.warn("User of: " + login + "failed to login");
 			return new ResponseEntity<>(0, HttpStatus.NOT_FOUND);
 		}
-		System.out.println(" user found! ");
+		
+		LOGGER.info(u.getUsername() + " has logged in");
 		//Return user id
 		return new ResponseEntity<>(u.getUser_id(), HttpStatus.OK);
 	}
 	
 	/*
-	 * GET user list by firstname*/
+	 * GET a list of users with firstname containing ?
+	 * Called when user searches in the navbar
+	 * */
 	@GetMapping("/searchFirstnames")
 	public List<User> searchUsersByFirstname(@RequestParam(name="firstname") String fn)
 	{
@@ -133,6 +153,10 @@ public class UserController {
 		return userList;
 	}
 	
+	/*
+	 * Get a list of users with username containing ?
+	 * Called when user searches in the navbar
+	 * */
 	@GetMapping("/searchUsernames")
 	public List<User> searchUsersByUsernames(@RequestParam(name="username") String un)
 	{
@@ -142,6 +166,10 @@ public class UserController {
 		return userList;
 	}
 	
+	/*
+	 * Return a user by Id
+	 * Called when attempting to view some ones profile
+	 * */
 	@GetMapping("/{username}")
 	public ResponseEntity<List<Post>> getUserByUserName(@PathVariable("username") String un)
 	{
@@ -155,12 +183,34 @@ public class UserController {
 		return new ResponseEntity<>(uService.getUserByUsername(un).getPosts(), HttpStatus.OK);
 	}
 	
+	/*
+	 * Return a user by Id
+	 * Called when attempting to view some ones profile
+	 * */
+	@GetMapping()
+	public ResponseEntity<User> getUserByUserId(@RequestParam(name="id") int id)
+	{
+		User u = uService.getUserById(id);
+		if (u == null)
+		{
+			return new ResponseEntity<>(null, HttpStatus.NOT_FOUND);
+		}
+		return new ResponseEntity<>(u, HttpStatus.OK);
+	}
+	
+	/*
+	 * Request to delete a user
+	 * */
 	@DeleteMapping("/{username}")
 	public ResponseEntity<String> deleteUser(@PathVariable("username") String un)
 	{
+		LOGGER.warn("Attempting to delete: " + un);
 		User u = uService.getUserByUsername(un);
 		
-		uService.deleteUser(u);
+		if(u != null)
+		{
+			uService.deleteUser(u);
+		}
 		
 		return new ResponseEntity<>("User Deleted", HttpStatus.GONE);
 	}
